working des loading

+++ PAPAS/src/main/java/lt/pap/dao/TofEngineRepository.java
@@ -4,7 +4,7 @@ import lt.pap.model.TofEngine;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TofEngineRepository extends JpaRepository<TofEngine, Long>
+public interface TofEngineRepository extends JpaRepository<TofEngine, Integer>
 {
     
 }
